requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
dnspython==2.1.0
emoji==1.5.0
pymongo==3.12.1
python-telegram-bot==13.7
pytz==2021.3
six==1.16.0
tornado==6.1
tzdata==2021.2.post0
tzlocal==3.0
